% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_MOTTE_forest.R
\name{build_MOTTE_forest}
\alias{build_MOTTE_forest}
\title{buildForest}
\usage{
build_MOTTE_forest(
  x.b,
  x.e,
  treat,
  y.b,
  y.e,
  nsplits = NULL,
  nodesize = 2 * (ncol(x.b) + 1),
  left.out = 0.1,
  ntree = ifelse(is.null(nsplits), 1, 200),
  nCore = ifelse(is.null(nsplits), 1, detectCores() - 1)
)
}
\arguments{
\item{x.b}{Before treatment covariates, a n by p matrix}

\item{x.e}{After treatment covariates, a n by p matrix}

\item{treat}{Treatment received, a n by 1 vector}

\item{y.b}{Before treatment outcomes, a n by q matrix}

\item{y.e}{After treatment outcomes, a n by q matrix
#@param method Method to use when choosing split value. Two options: "Exhaust" and "Random"}

\item{nsplits}{The number of split condidate want to examine when constructing split rule}

\item{nodesize}{Parameter to control the node size. When the number of observations in Node smaller than nodesize, stop splitting}

\item{left.out}{left.out is ensure at least left.out*2 sample for either treated or untreated sample in the group}

\item{ntree}{Number of trees want to construct. By default it is 1; however, when Random method used, recommand setting it as 200}

\item{nCore}{The number of cores use for forest contruction when doing parallel computation}
}
\value{
a list of data.tree. Even when only one tree construct, it is a list containing the single tree
}
\description{
The function to fitting tree/trees
}
\examples{
#' set.seed(1)
B <- create.B(10)
Z <- create.Z(10, 3)

tmp.dat <- sim_MOTTE_data( n.train = 500, n.test = 200,
p = 10, q = 3, ratio = 0.5,
B = B, Z = Z)

train.dat <- tmp.dat$train

x.b <- scale(train.dat$x.b, center = FALSE, scale = TRUE)
x.e <- scale(train.dat$x.e, center = FALSE, scale = TRUE)
y.b <- scale(train.dat$y.b, center = FALSE, scale = TRUE)
y.e <- scale(train.dat$y.e, center = FALSE, scale = TRUE)
treat <- train.dat$trt

# with(train.dat,
    build_MOTTE_forest(x.b, x.e, treat, y.b, y.e)
#)
}
