% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_MOTTE_tree_y.R
\name{build_MOTTE_tree_CO}
\alias{build_MOTTE_tree_CO}
\title{Fitting MOTTE Tree in the MOTTE.RF}
\usage{
build_MOTTE_tree_CO(
  x.b,
  x.e.1,
  x.e.2,
  y.b,
  y.e.1,
  y.e.2,
  nodesize,
  nsplits,
  left.out
)
}
\arguments{
\item{x.b}{Pre-treatment covariates, i.e. microbiomes}

\item{y.b}{Pre-treatment response, i.e. biomarkers}

\item{nodesize}{An integer value. The threshold that control the maximum size of a node}

\item{nsplits}{A numeric value, the number of maximum splits}

\item{left.out}{left.out is ensure at least left.out*2 sample for either treated or untreated sample in the group}

\item{x.e}{Post-treatment covariates, i.e. microbiomes}

\item{treat}{A vector of binary value, the arm of treatment}

\item{y.e}{Post-treatment response, i.e. biomarkers}

\item{seed}{a seed number to generate the random subsets of split candidates. Doesn't work when applying nsplit==NULL}
}
\value{
A data.tree object, node
}
\description{
Fitting MOTTE Tree in the MOTTE.RF
}
\examples{
set.seed(1)
B <- create.B(10)
Z <- create.Z(10, 3)

tmp.dat <- sim_MOTTE_data( n.train = 500, n.test = 200,
p = 10, q = 3, ratio = 0.5,
B = B, Z = Z)

train.dat <- tmp.dat$train

x.b <- scale(train.dat$x.b, center = FALSE, scale = TRUE)
x.e <- scale(train.dat$x.e, center = FALSE, scale = TRUE)
y.b <- scale(train.dat$y.b, center = FALSE, scale = TRUE)
y.e <- scale(train.dat$y.e, center = FALSE, scale = TRUE)
treat <- train.dat$trt
#with(train.dat,
    build_MOTTE_tree(x.b, x.e, factor(treat), y.b, y.e,
                     nodesize=30, nsplits=NULL, left.out = 0.1)
 #)

}
