% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{calcTrtDiff}
\alias{calcTrtDiff}
\title{Title}
\usage{
calcTrtDiff(forest, x.b)
}
\arguments{
\item{forest}{A MOTTE.RF object}

\item{x.b}{a data matrix for the testing data}
}
\value{
A data matrix that contains the treatment difference
}
\description{
Title
}
