% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MOTTE_data.R
\name{create.B1}
\alias{create.B1}
\title{Title
#TODO: improve the documentation}
\usage{
create.B1(p)
}
\arguments{
\item{p}{The dimension of the design matrix}
}
\value{
A matrix
}
\description{
Title
#TODO: improve the documentation
}
\examples{
create.B1(10)
}
