% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MOTTE_data_cross_over.R
\name{create.B_CO}
\alias{create.B_CO}
\title{Title
#TODO: improve the documentation}
\usage{
create.B_CO(p)
}
\arguments{
\item{p}{The dimension of the design matrix}
}
\value{
A matrix
}
\description{
Title
#TODO: improve the documentation
}
\examples{
create.B(10)
}
