% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MOTTE_data_cross_over.R
\name{create.Z_CO}
\alias{create.Z_CO}
\title{Title
#TODO: improve the documentation}
\usage{
create.Z_CO(p, q)
}
\arguments{
\item{p}{The number of rows}

\item{q}{The number of columns}
}
\value{
A design matrix
}
\description{
Title
#TODO: improve the documentation
}
\examples{
create.Z(10,3)
}
