% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{findMaxIndex}
\alias{findMaxIndex}
\title{Didn't find this function in tree ir forest yet. DOn't know if this is necessary}
\usage{
findMaxIndex(x, epsilon)
}
\arguments{
\item{x}{is the matrix}

\item{epsilon}{some threshold}
}
\value{

}
\description{
Didn't find this function in tree ir forest yet. DOn't know if this is necessary
}
