% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{MOTTE_feature_importance}
\alias{MOTTE_feature_importance}
\title{Calcualte feature importance}
\usage{
MOTTE_feature_importance(forest)
}
\arguments{
\item{forest}{}
}
\value{

}
\description{
Calcualte feature importance
}
