% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pop.var}
\alias{pop.var}
\title{Population variance calculation}
\usage{
pop.var(x)
}
\arguments{
\item{x}{a numeric vaector}
}
\value{
a numeric value
}
\description{
Population variance calculation
}
\examples{
pop.var(c(1,2,3))
}
