% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictResult}
\alias{predictResult}
\title{Predicting post-treatment responses}
\usage{
predictResult(tree.list, x.b)
}
\arguments{
\item{tree.list}{MOTTE forest}

\item{x.b}{a matrix, contain all observation that requires a prediction}
}
\value{
Predicted post-treatment responses for both treatment arm
}
\description{
Predicting post-treatment responses
}
