% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict_CO.R
\name{recommendResult.single}
\alias{recommendResult.single}
\title{Create predicted post-treatment responses for a single obervation}
\usage{
recommendResult.single(tree.list, x.b)

recommendResult.single(tree.list, x.b)
}
\arguments{
\item{tree.list}{MOTTE.RF object}

\item{x.b}{a vector of pre-treatment covariates for one observation}
}
\value{
a list contain the post-treament response mean for both treatment arms

a list contain the post-treament response mean for both treatment arms
}
\description{
Create predicted post-treatment responses for a single obervation

Create predicted post-treatment responses for a single obervation
}
