% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MOTTE_data.R
\name{sim_MOTTE_data}
\alias{sim_MOTTE_data}
\title{Simulate MOTTE formated data}
\usage{
sim_MOTTE_data(n.train, n.test, p, q, pi, seed = 1)
}
\arguments{
\item{n.train}{An integer value, sample size for training data}

\item{n.test}{An integer value, sample size for testing data}

\item{p}{An integer value, the dimensionality of covaraites (X)}

\item{q}{An integer value, the dimensionality of responses (Y)}

\item{pi}{An fraction value, the ratio between treatment groups}

\item{seed}{An integer value, set the seed number}
}
\value{
A nested list that contain training data and testing data
}
\description{
Simulate MOTTE formated data
}
\examples{
sim_MOTTE_data( n.train = 500, n.test = 200,
p = 10, q = 3, pi = 0.5)
}
