% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_CO.R
\name{traverseForest_CO}
\alias{traverseForest_CO}
\title{Traverse Forest
 A wrapper function using traverseTree to traverse the forest}
\usage{
traverseForest_CO(forest, x.b)
}
\arguments{
\item{forest}{A list of trees}

\item{x.b}{a single row vector contains the baseline variates of one observation}
}
\value{
A list of list containing OUTCOME and TREATMENT for each tree
}
\description{
Traverse Forest
 A wrapper function using traverseTree to traverse the forest
}
