% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_MOTTE_tree.R
\name{build_MOTTE_tree}
\alias{build_MOTTE_tree}
\title{Fitting MOTTE Tree in the MOTTE.RF}
\usage{
build_MOTTE_tree(x.b, x.e, treat, y.b, y.e, nodesize, nsplits, left.out)
}
\arguments{
\item{x.b}{Pre-treatment covariates, i.e. microbiomes}

\item{x.e}{Post-treatment covariates, i.e. microbiomes}

\item{treat}{A vector of binary value, the arm of treatment}

\item{y.b}{Pre-treatment response, i.e. biomarkers}

\item{y.e}{Post-treatment response, i.e. biomarkers}

\item{nodesize}{An integer value. The threshold that control the maximum size of a node}

\item{nsplits}{A numeric value, the number of maximum splits}

\item{left.out}{left.out is ensure at least left.out*2 sample for either treated or untreated sample in the group}
}
\value{
A data.tree object, node
}
\description{
Fitting MOTTE Tree in the MOTTE.RF
}
\examples{
# TODO: Add an example
}
