% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MOTTE_data.R, R/sim_MOTTE_data_cross_over.R
\name{create.B}
\alias{create.B}
\title{Title
#TODO: improve the documentation}
\usage{
create.B(p)

create.B(p)
}
\arguments{
\item{p}{The dimension of the design matrix}
}
\value{
A matrix

A matrix
}
\description{
Title
#TODO: improve the documentation

Title
#TODO: improve the documentation
}
\examples{
create.B(10)
create.B(10)
}
