% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is.between}
\alias{is.between}
\title{Checking if elements of a vector belongs to a exclusive interval from min to max}
\usage{
is.between(x, min, max)
}
\arguments{
\item{x}{A numeric vector/scalar}

\item{min}{the lower bound of the exclusive interval}

\item{max}{the upper bound of the exclusive interval}
}
\value{
True or False for valid comparison result. If \code{min} > \code{max}, the function return 0.

#@export
#@examples
#is.between(3, 1, 6)  #TRUE
#is.between(7, 1, 6)  #FALSE
#is.between(5, 4, 1)  #0

#is.between(c(1,3,2,7), 1, 6)
}
\description{
Checking if elements of a vector belongs to a exclusive interval from min to max
}
