% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is.between}
\alias{is.between}
\title{Checking if elements of x belongs to a exclusive interval from min to max}
\usage{
is.between(x, min, max)
}
\arguments{
\item{x}{The elemnt want to check. Must be an integer}

\item{min}{the lower bound}

\item{max}{the upper bound}
}
\value{
True or False
}
\description{
Checking if elements of x belongs to a exclusive interval from min to max
}
\examples{
a <- 5
is.between(a, 1, 6)
}
