% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict_CO.R
\name{recommendResult}
\alias{recommendResult}
\title{Recommend treatment based on pre-treatment covariates and a weight}
\usage{
recommendResult(tree.list, x.b, w)

recommendResult(tree.list, x.b, w)
}
\arguments{
\item{tree.list}{MOTTE forest}

\item{x.b}{a maitrx, a matrix that contains pre-treatment covarites}

\item{w}{a numeric vector, indicating the direction of desired outcome}
}
\value{
a vector that contains the recommanded treatment level for each individual

a vector that contains the recommanded treatment level for each individual
}
\description{
Recommend treatment based on pre-treatment covariates and a weight

Recommend treatment based on pre-treatment covariates and a weight
}
